+++ R/upper_tier_region_ctry_lookup.R
@@ -1,4 +1,4 @@
-#' A lookup of upper-tier Local Authorities (EW) to their Region (if any) and Country
+#' (data) A lookup of upper-tier Local Authorities (EW) to their Region (if any) and Country
 #'
 #' @format A data frame with 174 rows and 6 variables:
 #' \describe{