+++ tests/testthat/test-build_api_query.R
@@ -3,7 +3,6 @@ test_that(
     expect_identical(
       build_api_query(
         ref = 4,
-        type = "census",
         within_level = "cauth20nm",
         within = "Greater Manchester",
         fields = c("lad20cd", "lad20nm", "cauth20cd", "cauth20nm")
@@ -22,7 +21,6 @@ test_that(
     expect_identical(
       build_api_query(
         ref = 10,
-        type = "census",
         within_level = "msoa11cd",
         within = c("E02006666", "E02006667", "E02006668"),
         fields = c("msoa11cd", "msoa11nm")
@@ -41,7 +39,6 @@ test_that(
     expect_identical(
       build_api_query(
         ref = 12,
-        type = "census",
         within_level = "lad20nm",
         within = c("Cheltenham", "Gloucester", "Stroud", "Cotswold", "Tewkesbury", "Forest of Dean"),
         fields = c("lad20cd", "lad20nm", "BNG_E", "BNG_N"),